--- gel-db-protocol/src/gen2/mod.rs
@@ -1,4 +1,0 @@
-mod macros;
-mod traits;
-
-pub use traits::*;
